Pass X-Nerve-Check-IP to configure_nerve
This is so that we send the IP of Kubernetes Pods to Nerve in this
header. That way a system level hacheck knows what IP to check against.
Note that I am also making us send the header in the case of marathon
tasks. This is so we can use the same hacheck for both marathon and
kubernetes tasks.